requirements.txt
google-generativeai>=0.8.3
python-dotenv>=1.0.1
protobuf>=5.26.1,<6.0dev
flask>=2.0.0